Subscribe to webhooks for repo update/delete
Take repo:update and repo:delete events into account so that wikis are updated
accordingly. repo:post-update events now use repo ids as the callback so that
they are "name independent". Repos that are renamed are simply updated in the
db, whilst repos that are deleted also delete their corresponding repos.