Fix double-tilde bug in the delete wiki page
The routes for `user_content` in mansrht/blueprints/html.py include a
tilde ('~') in them, which is required for the route to function
properly.  However, .canonical_name already includes a tilde[a].  This
causes in a bug where any usage of user_content with the
owner.canonical_name field will result in a generated url where the
username is prefixed with two tildes ("~~nytpu") rather than one.

This replaces the usages of owner.canonical_name with owner.username
which is identical to .canonical_name (see link [a]) except that it does
not have a leading tilde.

[a]: https://git.sr.ht/~sircmpwn/core.sr.ht/tree/99a8168de436c576ec98242d255a71a2b94c9658/item/srht/oauth/user.py#L37