Add links back to wiki root on wiki pages
This adds links back to the wiki root index in two places: on the
pseudo-breadcrumbs with the username and wiki name, and on the "article"
tab.  This makes it possible to navigate back to the wiki index from any
wiki page, whereas before it was impossible.  The new links also fixes
the most counterintuitive link on the page, as the "article" tab linking
back to the site root (/) is very strange and unexpected.