create: fix logic error with existing refs
The "existing" branch in the template is used on the repo select screen
to filter out repositories which already have a wiki created for them.
This change adds a short-circuit which avoids the check in other
situations.

While I'm here I also tweaked the appearance of this page a bit.