content.html: show wiki branch information
Since wikis can be backed by arbitrary branches, the "Clone this wiki"
part of pages such as https://man.sr.ht/~rjarry/aerc, that only gives
clone URLs, can be ambiguous.

While it is possible to find the branch information when you know the
basics about SourceHut wikis, it's an unnecessary hurle for "new folks"
to contribute.

This patch adds the branch information to "Clone this wiki".

Signed-off-by: Simon Martin <simon@nasilyan.com>