blueprints/html: fix redirect handling
The redirect handling here is pretty tricky. The correct behavior should
be:

- Always require a trailing slash if asking for a tree object
- EXCEPT for the root of the root wiki (i.e. "https://man.sr.ht")

There are some things to watch our for. The `path` variable is already
processed, and is not suitable for testing if the requested URL ends
with a slash. Hence, test `request.path` instead. Also, make sure to
check if this is the root wiki, and only suppress sending the redirect
if it's the root wiki _and_ the root path.